rgw: add bucket chown entrypoint
Added an admin rest entrypoint for changing the owner of a bucket.
The call is mapped over the existing: RGWHandler_Bucket::op_put().
The query string must contain the parameter: owner in order to trigger a chown.

Example:

admin/bucket?owner&uid=john&bucket=foo

This invocation changes the bucket foo's owner with the user john.
All the objects inside the bucket are updated according to the new owner
as well.

The user calling chown entrypoint must have users capability set with write permission.
On success, the operation returns a 200 HTTP code with a body summarizing the call's details; example:

{
	old_user_id : "jack",
	new_user_id : "john",
	object_progression : [
		{
			processed_objects : 100,
			processing_object : "obj_101"
		},
		{
			processed_objects : 200,
			processing_object : "obj_201"
		}
	],
	processed_objects : 201
}

The object_progression array is incrementally sent to the client as the
operation proceeds.
This behavior prevents a client to timeout when the bucket contains a
significant amount of objects.

Fixes: https://tracker.ceph.com/issues/57569
Signed-off-by: Giuseppe Baccini <giuseppe.baccini@suse.com>

working on chown with flusher

Signed-off-by: Giuseppe Baccini <giuseppe.baccini@suse.com>

working on chown with flusher 2

Signed-off-by: Giuseppe Baccini <giuseppe.baccini@suse.com>