qa/suites/crimson-rados: remove cache and dedup related tests
It's an accident that some of these pass -- the tests provide some
coverage for inconsistencies that can arise, but don't really validate
that the objects actually get moved between tiers.

It's going to be some time before we implement cache or dedup in
crimson, and we'll probably want to disable the related commands for
pools that can map to crimson osds to prevent accidents.

Signed-off-by: Samuel Just <sjust@redhat.com>