doc/radosgw: add push_endpoint for rabbitmq
This commit directs users of rabbitmq to use "push_endpoint"
(with an underscore) instead of "push-endpoint" (with a hy-
phen). This commit adds a note that contains such a direct-
ive. It does not alter the examples already present in the
text.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit ddf8295438b14ada04d8bb5dfc23402bedc937b8)