mgr/cephadm: save host cache data after scheduling daemon action
Otherwise, the actions could be lost on a mgr failover
despite us including shceduled actions as part of the data
we store in the config-key entry for the host.

This also makes it impossible to redeploy the active mgr
with a new image in one command.

Fixes: https://tracker.ceph.com/issues/57884

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 2ff3ae71fac22f9267451442b9e43c49375e40e3)