rgw: set keys from from master zone on admin api user create
When users were being created via the admin_api
on a secondary zone the keys returned in the json
output are not correct.

This commit ensures the proper keys are set to be
returned to the user via the json output.

fixes: https://tracker.ceph.com/issues/57724

Signed-off-by: Ali Maredia <amaredia@redhat.com>
(cherry picked from commit e3a26f77c090076eb76bf67af659ce88af9eee8e)