doc/cephadm: update install.rst
This commit addresses https://github.com/ceph/ceph/pull/43250,
which has remained unmerged for thirteen months. The issues
raised in that PR are addressed here.

This commit contributes to the simplification of the procedure
that explains how to install cephadm. More remains to be done
on this page, but the old PR had to be dealt with first.

Signed-off-by: Zac Dover <zac.dover@gmail.com>