test/librbd/fsx: add rbd-wnbd tests
rbd-wnbd can be used to map rbd images as Windows block devices.

This change will allow exercising rbd-wnbd and the wnbd driver
as opposed to using librbd directly.

One thing to note here is that read and write operations must be
512B aligned. Also, wnbd does not currently support "compare
and write".

For this reason, we'll stick to the basic operations for the time
being.

Note that we're renaming krbd_read/krbd_write to generic_pread
and generic_pwrite, reusing those functions for wnbd operations.
We're also removing the duplicated ggate read/write operations,
using the generic functions instead.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>