osd: don't lose information when stringifying pg state into 'unknown'
The reason behind this patch is direct attention to the fact
we lose a pretty usueful information for troubleshooting when
a PG state is being converted into a string representation.

If there is no available human-readable state name, "unkown"
(without any extra details) is used. This means multiple
different states can be strigified to the same string which
consitutes the data loss.

This patch alters the unknown state by appending the exact
state value in hex. The format becomes "unknown(%x)".

Signed-off-by: Radosław Zarzyński <rzarzyns@redhat.com>