ceph-crash: Use auth_names when pinging cluster at startup
The code to ping the cluster with a key on startup doesn't specify
a client name, which means it may not be using the correct key.
This becomes obvious when running inside a container, which will
have the client.crash.$hostname keyring, but not the admin keyring.

This commit is only necessary for the main branch, as the original
ping on startup was never backported anywhere else.

Fixes: 84c4562f7bff9ca1018fd1975657144e62d6b93d
Signed-off-by: Tim Serong <tserong@suse.com>