rgw: Add log types to dencoder
When debugging an issue in RGW multisite replication, we discovered that we had
no way to deserialise data stored in the logs (in particular the datalog) when
attempting to look inside the RADOS objects and work out what's going on. This
commit adds the types to ceph-dencoder, to facilitate future efforts at
debugging. As a bonus, this also allows serialisation for these types to be
tested.

Adds the following types to ceph-dencoder:

- RGWMetadataLogData
- rgw_data_change
- cls_log_entry

Signed-off-by: Tom Coldrick <tcoldrick@bloomberg.com>