crimson/osd: add formatter for std::optional<>
in `PG::do_update_log_missing()`, `std::optional<eversion_t>` is printed
using {fmt}, but {fmt} does not support formatting `std::optional<>`. so
we need to swing our own formatter. we can promote this formatter to
a more common place if it is proved to be useful otherwhere in the
source tree.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>