doc/install: update "Official Releases" sources
Remove "Docker Hub" from the list of locations from which Ceph can be
acquired.

This commit represents a second attempt to get
https://github.com/ceph/ceph/pull/49034 past the Jenkins CI tests, and
that PR was itself an attempt to incorporate the change in
https://github.com/ceph/ceph/pull/44310, which could not be merged
because the commit message associated with that PR did not begin with
the string "doc", which meant that the "Signed-off-by" check failed.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit 22088630fd89319ca1526d5dc86cbc14a889b1a3)