*: include fmt/ostream.h if FMT_VERSION >= 9
before this change, we assume that `fmt/ostream.h` is included, and
reference `fmt::ostream_formatter`. this breaks the build if the tree
is compiled on {fmt} v9 and up.

after this change, `fmt/ostream.h` is conditionally included. as the
compiling unit is supposed to include fmt headers before including
headers in this project, we can tell if fmt/ostream.h is needed or not
by checking the FMT_VERSION macro.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>