crimson/os: specialize fmt::formatter<> for StagedIterator
StagedIterator is a nested class of stage_t, which is in turn a template
class. it would be impossible to partial specialize fmt::formatter<> for
a nested class of a template class. to workaround this, we specialize
fmt::formatter<> for a type which has `do_format_to()` method, and
define this method for StagedIterator.

since seastar::logger is the only user of the operator<<() operator,
the original operator<<() operator is replaced with this new helper
method.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>