dokan: Made file/dir access mode configurable
By default, when mounting a filesystem using ceph-dokan, the file
and directory access modes were hardcoded as 755.

Now, both the file and directory access modes are configurable
using the --file-mode and --dir-mode optargs, accepting values
ranging from 001 to 777.

If no value is specified, the default value of 755 will be used.

Signed-off-by: Stefan Chivu <schivu@cloudbasesolutions.com>