RGW - Zipper - Rename rgw::sal::Store to rgw::sal::Driver
All along, we've had an overload of the term "store".  It was the base
class of the entrypoint, and it was the lowest layer in a stack.  This
renames the base class to Driver.  So, the primary elements of zipper
that to work are now Drivers, and they come in two different flavors:
Filters live in the middle of the stack, and do not need to implement
every API, but pass through instead.  And Stores live a the bottom of
the stack, and must implement the entire API.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>