ceph-volume: fix a bug in lsblk_all()
Rook has a specific use case where devices are copied in /mnt

If the basename (in /mnt) is different from the original device name,
then the current logic can't match it.

The idea is to append the device to the `lsblk` command and return
the result.

Fixes: https://tracker.ceph.com/issues/58137

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>
(cherry picked from commit dce4d469f66124babe97dd3651f67d55e805088c)