os/bluestore: DO NOT MAKE RELEASE WITH THIS PR
This PR is for discovering what set of problems is caused by possible
error of incorrect reading of BlueFS log.

The assumption is that the following might be a problem:
1) BlueFS log is always written in buffered=false mode
2) BlueFS log is read in bluefs_buffered_io mode

If we detect inconsistency, we allow code to run.
Then we append message "DELAYED CRITICAL FAILURE" to sigsegv / assert() handlers,
if it eventually happens.
Or we abort after 100s.

Signed-off-by: Adam Kupczyk <akupczyk@redhat.com>