osd: fix couples of bugs with similar error log of bug 52657. log errors like MOSDPGLog::encode_payload(uint64_t): Assertion `HAVE_FEATURE(features, SERVER_NAUTILUS)'
calling AsyncMessenger::connect_to is asynchronous and just set in state of START_CONNECT; a sequence of handshakes frame(including server identity frame) is pending to be exchanged before the connection is in ready state. The race condition is then sending message by the connection should not be allowed. it should wait for readiness if a connection be used instantly after its creation.

Since the issue is hard to reproduce, only with some sort of race condition, it would be better to bring up a live cluster in lab with the commit to see how it works for us.

this should fix the following similar issues:
https://tracker.ceph.com/issues/57845
https://tracker.ceph.com/issues/52657
https://tracker.ceph.com/issues/52694
https://tracker.ceph.com/issues/57845
https://tracker.ceph.com/issues/53685

Signed-off-by: Ben Gao <bengao168@msn.com>