mds: remove mdsdir inode from scrubbing_map if scrub stack is empty
Even after a successful scrub, mdsdir inode is not cleared from scrubbing_map,
reason for this is not very clear to me yet. Interesting thing to notice here is
that logs do say that this inode went through ScrubStack::dequeue which removes
the inodes from scrub_stack but still this issue has been existent throughout my
teuthologys runs. Therefore as a cure, I think it is safe to say that when the
scrub_stack is empty and the mdsdir inode is not purging, we can just remove it's
entry from scrubbing_map. Rationale being: purge is followed by scrub(atleast this
has happened in all of my runs), therefore saying that it is safe to remove mdsdir
inode when mdsdir is not purging and also it's not in scrub stack should be a valid
argument.

Fixes: https://tracker.ceph.com/issues/53724
Signed-off-by: Dhairya Parmar <dparmar@redhat.com>