pybind/mgr: do not test with py37
as we always test with ubuntu jammy, which does not provide python3.7:

py37: skipped because could not find python interpreter with spec(s): py37

so there is no point testing with python3.7.

also, in tox v4, it is not able to handle "key = value" anymore, where
value has newlines in it. so we need to find a better way passing
command line options to the test command.

this change partially reverts c797ace324d951f06c8e35eb3b87f192d21fd7a5

Signed-off-by: Kefu Chai <tchaikov@gmail.com>