mgr/pybind: fix mypy arg parsing
On the new tox version it is treating each line as a new command
so it will do something like "mypy --config-file=../../mypy.ini"
as one command and then "-m balancer" as a totally separate command.
The first one immediately fails as it doesn't include any modules
to test. Adding backslashes to the ends of the lines gets it to
handle the lines as one long command

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 8e5840ff2c5815897e3eaeb136927971ad6f5384)