cmake: pass cflags using user-config.jam
it is "b2 install" which builds the target. "stage" defines "install"
rule by copying the target to a single location. so we need to pass
them to the "install" rule. now that we already have "user-config.jam"
defining the used toolset like gcc, we can just use it for passing
the cflags and cxxflags.

in this change, the cflags set with BOOST_CXXFLAGS are persisted in
"user-config.jam". so they won't get lost when building Boost using
"b2 install".

also, strictly speaking the flags of "-fPIC -w" are not
CXXFLAGS, they are compile flags. so including them in "compileflags"
is more correct.

Signed-off-by: Kefu Chai <kchai@redhat.com>