crimson/osd: include fmt/ranges.h
so we are able to format ss.clone_snaps, which is a
std::map<snapid_t, std::vector<snapid_t>>. before fmtlib v9,
fmtlib is able to fall back to the operator<<() to format the the types
with the operator<<() defined. but after fmtlib v9, we need to
explicitly define them for accessing the specialized formatter.

in this change, fmt/ranges.h is included so we can access the formatter
for std::map and std::vector.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>