rgw: add logging to radosgw-admin bi list
On a nautilus version we ran into an issue where bi list was not
listing out all entries and also not reporting any errors. Even after
turning up logging it was impossible to follow the code paths. This
adds level 20 logging in order to follow the code paths even if no
error is generated, so we can better handle such situations in the
future.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>