tools/rados: Modify rados bench to setup client QoS.
Add a new option called "with-qos" to rados bench that when specified
reads the client specific reservation, weight and limit config options,
creates and sets the QoS profile using the new librados QoS APIs.

The QoS parameters are passed to the dmclock scheduler using the QoS
profile for the client.

With these changes, any client using rados bench will be able to
specify it's QoS parameters by just specifying the "with-qos" option
as part of the rados bench command.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>