mds: record and dump last tid for trimming completed requests (or flushes)
CephFS clients include `oldest_tid` as part of the client request
to the MDS. This field is the tid of the oldest incomplete mds
request (excluding setfilelock request). The MDS uses this to
trim completed requests (and flushes). In one case, the ceph
cluster had an extremely high completed requests count, meaning,
for some reason the client was not advancing its `oldest_tid`
field, although, the MDS had successfully "safe replied" the
request back to the client.

This change adds a debug aid for recording and dumping this
field. It might be possible to fetch this from clients (if
not, we should add that!), but it makes sense to have this
information available from the MDS.

Partially-Fixes: http://tracker.ceph.com/issues/57985
Signed-off-by: Venky Shankar <vshankar@redhat.com>
(cherry picked from commit d01acd531a0694514f7e7e11087ce7a9d16e94d9)

 Conflicts:
	doc/cephfs/health-messages.rst

Minor conflict w.r.t. line breaks in the document.