rados: Set single compression algorithm as a default value in ms_osd_compression_algorithm instead of list of algorithms
The default value for the compression parameter "ms_osd_compression_algorithm" is assigned to four different algorithms. This is a bit confusing as we donot know which algorithm is in use by the OSDs. The OSDs choose the preferred algorithm based on mutual understanding between OSDs. The proposed solution to this problem was to set "snappy" as default algorithm and provide flexibility to user to change the algorithm of their choice.

Fixes: https://tracker.ceph.com/issues/58410

Signed-off-by: shreyanshjain7174 <ssanchet@redhat.com>